% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox}
\alias{boxcox}
\title{This function applies a box cox transform to a vector.
Note that all entries of the vector must be strictly greater than 0.}
\usage{
boxcox(x, lambda, plot_it = FALSE)
}
\arguments{
\item{x}{The vector to apply the box cox transform to}

\item{lambda}{The parameter used to perform the transform}

\item{plot_it}{Display a plot of the output of the transform versus x, if set to TRUE}
}
\value{
This function performs returns the box cox transform on x, using parameter lambda
}
\description{
This function applies a box cox transform to a vector.
Note that all entries of the vector must be strictly greater than 0.
}
